error handling
I added some error handling. For example, if there aren't any entries in
the database an undefined variable warning is displayed. I declared the
$e variable at the top of the else statement as an empty string. This
works because the loop will create $e as an array but there are no
entries, there is no array and therefor no variable.  The check at the
bottom of the file for any entries determines if the $e variable is an
array or not. If there weren't any entries it doesn't create the
variable whether it's an array or not. By declaring it as a string the
function handles it whether there are entries or not.